pyproject.toml
[project]
name = "dont-be-late"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client",
    "google-auth-oauthlib",
    "PyQt6",
    "pyobjc-core",
    "pyobjc-framework-Cocoa",
    "pyobjc-framework-Quartz",
    "pyyaml",
]

[project.scripts]
dont-be-late = "dont_be_late.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dont_be_late"]

[tool.uv]
package = true
